+++ extras/tx-indexer-testsuite/build.nix
@@ -59,7 +59,7 @@
         };
         settings.processes = {
           tests = {
-            command = "cargo test";
+            command = "${pkgs.cargo}/bin/cargo test";
             depends_on = {
               cardano_devnet.condition = "process_healthy";
               ogmios.condition = "process_healthy";